+++ libraries/resource_providers/api600/c7000/server_hardware_provider.rb
@@ -14,6 +14,10 @@ module OneviewCookbook
     module C7000
       # ServerHardware API600 C7000 provider
       class ServerHardwareProvider < API500::C7000::ServerHardwareProvider
+        def add_multiple_servers
+          multiple_servers = resource_named(:ServerHardware).new(@item.client, @item.data)
+          multiple_servers.add_multiple_servers
+        end
       end
     end
   end