+++ spec/unit/resources/matchers_spec.rb
@@ -220,6 +220,7 @@ describe 'oneview_test::default' do
 
     # oneview_server_hardware
     expect(chef_run).to_not add_oneview_server_hardware_if_missing('')
+    expect(chef_run).to_not add_multiple_servers_oneview_server_hardware('')
     expect(chef_run).to_not remove_oneview_server_hardware('')
     expect(chef_run).to_not refresh_oneview_server_hardware('')
     expect(chef_run).to_not set_oneview_server_hardware_power_state('')
@@ -288,6 +289,7 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_volume_if_missing('')
     expect(chef_run).to_not delete_oneview_volume('')
     expect(chef_run).to_not create_oneview_volume_snapshot('')
+    expect(chef_run).to_not repair_oneview_volume('')
     expect(chef_run).to_not delete_oneview_volume_snapshot('')
     expect(chef_run).to_not add_oneview_volume_if_missing('')
     expect(chef_run).to_not create_oneview_volume_from_snapshot('')